requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
python-chess==1.999
python-multipart==0.0.6
requests==2.31.0
